+++ src/game/Object/Player.cpp
@@ -7543,185 +7543,6 @@ Pet* Player::GetMiniPet() const
 }
 
 /**
- * @brief Sends the controlled pet's spell bar and cooldown state to the client.
- */
-void Player::PetSpellInitialize()
-{
-    Pet* pet = GetPet();
-
-    if (!pet)
-    {
-        return;
-    }
-
-    DEBUG_LOG("Pet Spells Groups");
-
-    CharmInfo* charmInfo = pet->GetCharmInfo();
-
-    WorldPacket data(SMSG_PET_SPELLS, 8 + 4 + 1 + 1 + 2 + 4 * MAX_UNIT_ACTION_BAR_INDEX + 1 + 1);
-    data << pet->GetObjectGuid();
-    data << uint32(0);
-    data << uint8(charmInfo->GetReactState()) << uint8(charmInfo->GetCommandState()) << uint16(0);
-
-    // action bar loop
-    charmInfo->BuildActionBar(&data);
-
-    size_t spellsCountPos = data.wpos();
-
-    // spells count
-    uint8 addlist = 0;
-    data << uint8(addlist);                                 // placeholder
-
-    if (pet->IsPermanentPetFor(this))
-    {
-        // spells loop
-        for (PetSpellMap::const_iterator itr = pet->m_spells.begin(); itr != pet->m_spells.end(); ++itr)
-        {
-            if (itr->second.state == PETSPELL_REMOVED)
-            {
-                continue;
-            }
-
-            data << uint32(MAKE_UNIT_ACTION_BUTTON(itr->first, itr->second.active));
-            ++addlist;
-        }
-    }
-
-    data.put<uint8>(spellsCountPos, addlist);
-
-    uint8 cooldownsCount = pet->m_CreatureSpellCooldowns.size() + pet->m_CreatureCategoryCooldowns.size();
-    data << uint8(cooldownsCount);
-
-    time_t curTime = time(NULL);
-
-    for (CreatureSpellCooldowns::const_iterator itr = pet->m_CreatureSpellCooldowns.begin(); itr != pet->m_CreatureSpellCooldowns.end(); ++itr)
-    {
-        time_t cooldown = (itr->second > curTime) ? (itr->second - curTime) * IN_MILLISECONDS : 0;
-
-        data << uint16(itr->first);                         // spellid
-        data << uint16(0);                                  // spell category?
-        data << uint32(cooldown);                           // cooldown
-        data << uint32(0);                                  // category cooldown
-    }
-
-    for (CreatureSpellCooldowns::const_iterator itr = pet->m_CreatureCategoryCooldowns.begin(); itr != pet->m_CreatureCategoryCooldowns.end(); ++itr)
-    {
-        time_t cooldown = (itr->second > curTime) ? (itr->second - curTime) * IN_MILLISECONDS : 0;
-
-        data << uint16(itr->first);                         // spellid
-        data << uint16(0);                                  // spell category?
-        data << uint32(0);                                  // cooldown
-        data << uint32(cooldown);                           // category cooldown
-    }
-
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends the possessed unit's action bar state to the client.
- */
-void Player::PossessSpellInitialize()
-{
-    Unit* charm = GetCharm();
-
-    if (!charm)
-    {
-        return;
-    }
-
-    CharmInfo* charmInfo = charm->GetCharmInfo();
-
-    if (!charmInfo)
-    {
-        sLog.outError("Player::PossessSpellInitialize(): charm (GUID: %u TypeId: %u) has no charminfo!", charm->GetGUIDLow(), charm->GetTypeId());
-        return;
-    }
-
-    WorldPacket data(SMSG_PET_SPELLS, 8 + 4 + 4 + 4 * MAX_UNIT_ACTION_BAR_INDEX + 1 + 1);
-    data << charm->GetObjectGuid();
-    data << uint32(0);
-    data << uint32(0);
-
-    charmInfo->BuildActionBar(&data);
-
-    data << uint8(0);                                       // spells count
-    data << uint8(0);                                       // cooldowns count
-
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends the charmed unit's available actions and spells to the client.
- */
-void Player::CharmSpellInitialize()
-{
-    Unit* charm = GetCharm();
-
-    if (!charm)
-    {
-        return;
-    }
-
-    CharmInfo* charmInfo = charm->GetCharmInfo();
-    if (!charmInfo)
-    {
-        sLog.outError("Player::CharmSpellInitialize(): the player's charm (GUID: %u TypeId: %u) has no charminfo!", charm->GetGUIDLow(), charm->GetTypeId());
-        return;
-    }
-
-    uint8 addlist = 0;
-
-    if (charm->GetTypeId() != TYPEID_PLAYER)
-    {
-        CreatureInfo const* cinfo = ((Creature*)charm)->GetCreatureInfo();
-
-        if (cinfo && cinfo->CreatureType == CREATURE_TYPE_DEMON && getClass() == CLASS_WARLOCK)
-        {
-            for (uint32 i = 0; i < CREATURE_MAX_SPELLS; ++i)
-            {
-                if (charmInfo->GetCharmSpell(i)->GetAction())
-                {
-                    ++addlist;
-                }
-            }
-        }
-    }
-
-    WorldPacket data(SMSG_PET_SPELLS, 8 + 4 + 1 + 1 + 2 + 4 * MAX_UNIT_ACTION_BAR_INDEX + 1 + 4 * addlist + 1);
-    data << charm->GetObjectGuid();
-    data << uint32(0x00000000);
-
-    if (charm->GetTypeId() != TYPEID_PLAYER)
-    {
-        data << uint8(charmInfo->GetReactState()) << uint8(charmInfo->GetCommandState()) << uint16(0);
-    }
-    else
-    {
-        data << uint8(0) << uint8(0) << uint16(0); // TODO it is exactly the same as uint32(PetModeFlags) from SMSG_PET_MODE
-    }
-
-    charmInfo->BuildActionBar(&data);
-
-    data << uint8(addlist);
-
-    if (addlist)
-    {
-        for (uint32 i = 0; i < CREATURE_MAX_SPELLS; ++i)
-        {
-            CharmSpellEntry* cspell = charmInfo->GetCharmSpell(i);
-            if (cspell->GetAction())
-            {
-                data << uint32(cspell->packedData);
-            }
-        }
-    }
-
-    data << uint8(0);                                       // cooldowns count
-
-    GetSession()->SendPacket(&data);
-}
-
-/**
  * @brief Clears the pet action bar on the client.
  */
 void Player::RemovePetActionBar()
